split up components for better readability
