+++ packages/admin/admin/src/utils/getElementOuterWidth.ts
@@ -1,0 +1,2 @@
+export const getElementOuterWidth = (element: Element): number =>
+    element.clientWidth + parseFloat(getComputedStyle(element).marginLeft) + parseFloat(getComputedStyle(element).marginRight);
